+++ CHANGELOG.md
@@ -7,6 +7,8 @@
  - Added an option for GM's to force whispers to always be tagged as In Character, to play nice with the Tabbed Chatlog module.
  - Added a global, per client option to show a toast notification when they received a whisper.
  - Added a global, per client configurable option to play a sound when receiving a new whisper.
+ - Added a new one-time message upon a user signing in, introducing them to (Whisper) Sweet Nothings.
+ - Added a whisper history that filters based on the selected recipient(s).
 
 ## v1.2.2 - 2022-03-14
  - Added Japanese translation thanks to @besardida!